+++ code/src/App.js
@@ -1,13 +1,13 @@
 import React from 'react'
 import data from './data.json'
-import Album from './Album'
+import Album from './components/Album'
 
 
 export const App = () => {
 
 // AN ARRAY WITH ALL ALBUMS
 const albumArray = data.albums.items
-// console.log(albumArray)
+console.log(albumArray)
 
 // AN ARRAY WITH ARTISTS
 const allArtists = albumArray.map(album => album.artists)
@@ -19,6 +19,7 @@ const nameOfArtists = allArtists.map(artist => artist.map((item) => item.name))
 
 // URLS TO ARTISTS
 const artistURL = allArtists.map(artist => artist.map((item) => item.external_urls.spotify))
+console.log(artistURL)
   
 
   return (
@@ -34,8 +35,9 @@ const artistURL = allArtists.map(artist => artist.map((item) => item.external_ur
               <Album 
               image={album.images[0].url} 
               title={album.name} 
-              href={album.external_urls.spotify} 
-              artist={album.artists.map((item) => item.name)}
+              hrefAlbum={album.external_urls.spotify} 
+              artist={album.artists.map((item, index) => item.name)}
+              hrefArtist={album.artists.map((item) => item.external_urls.spotify)}
               key={album.name}/>
               )
           })
