+++ code/src/components/Album-image.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const AlbumImage = (prop) => {
+    return (
+          <img className="albumImage" id="albumImage" src={prop.image}/>   
+    )
+}
+
+export default AlbumImage
