+++ code/src/components/Icons.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const Icons = (prop) => {
+    return (
+        <div className="icons hidden" id="hidden">
+                    <img className="icon-heart hidden" src="/icons/heart.svg" alt="favorite"/>
+                    <img className="icon-play" src="/icons/play.svg" alt="play"/>
+                    <img className="icon-dots" src="/icons/dots.svg" alt="more options"/>
+                </div>
+    )
+}
+
+export default Icons
