+++ code/src/index.css
@@ -13,6 +13,10 @@ code {
     monospace;
 }
 
+a {
+  text-decoration: none;
+}
+
 /* HEADING */
 
 .heading {
